+++ tsp/launch.sh
@@ -1,6 +1,8 @@
+#!/bin/bash
+
 file_name="tsp-bf"
 clear
-rm $file_name.out
+[ -e $file_name.out ] && rm $file_name.out
 gcc -Wall -fopenmp -o $file_name.out $file_name.c
 [ $2 ] && a=$2  || a="default"
 printf "Size: %s and Seed: %s\n\n" $1 $a
